Downloads/Step2_Webアプリ/4Googlevisionapi/streamlitmoc/requirements.txt
streamlit==1.27.2
google-cloud-vision==3.4.5


